refactor: Move Menu button on home to component
